+++ app/(tabs)/(onboard)/nickname.tsx
@@ -3,7 +3,7 @@ import Text from '@/components/Text';
 import { TextSize } from '@/enums/TextSize';
 import View from '@/components/View';
 import Textfield from '@/components/Textfield';
-import Button from '@/components/Button';
+import Button from '@/components/ui/atoms/Button';
 import useOnboardStore from '@/stores/onboardStore';
 import { router } from 'expo-router';
