+++ app/(tabs)/(onboard)/signup.tsx
@@ -2,7 +2,7 @@ import {Keyboard, Pressable, StyleSheet, useColorScheme} from 'react-native';
 import Text from '@/components/Text';
 import { TextSize } from '@/enums/TextSize';
 import View from '@/components/View';
-import Button from '@/components/Button';
+import Button from '@/components/ui/atoms/Button';
 import Textfield from '@/components/Textfield';
 import { useEffect, useState } from 'react';
 import { RegexList } from '@/constants/RegexList';
